vbs: restore runtest.py to updated version
